Translate 4-digit ESCO to ISCO08

This function translates a vector of 4-digit ESCO codes to ISCO08 codes using the translation table stored in all_schema$esco_to_isco08.

Usage

esco_to_isco08(x, label = FALSE, to_factor = FALSE)

Arguments

x

A character vector of 4-digit ESCO codes.

label

A logical value indicating whether to return the labels of the translated ISCO08 codes (default is FALSE).

to_factor

A logical value indicating whether to return a factor instead of a character. The order of the labels is taken from the labels for ISCO08 found in all_labels (default is FALSE).

Value

A character vector of ISCO08 codes.

Details

This translation converts the ESCO classification of the European Commission to ISCO08. This translations uses ESCO v1.1.1 and was downloaded from the dataset https://esco.ec.europa.eu/en/use-esco/download.

For more information on this class schema, please check the references below:

The ESCO Classification: https://esco.ec.europa.eu/en/classification

Examples

esco_to_isco08(c("11401", "1101", "11301"), label = FALSE)
#> [1] NA     "1100" "1100" NA    

esco_to_isco08(c("11401", "1101", "11301"), label = TRUE)
#> [1] NA                                                    
#> [2] "'Chief executives, senior officials and legislators'"
#> [3] "'Chief executives, senior officials and legislators'"
#> [4] NA                                                    

esco_to_isco08(c("11401", "1101", "11301"), label = TRUE, to_factor = TRUE)
#> [1] <NA>                                                
#> [2] 'Chief executives, senior officials and legislators'
#> [3] 'Chief executives, senior officials and legislators'
#> [4] <NA>                                                
#> 582 Levels: 'Managers' < ...