Translates 2-digit ISCO08 to ESEG.

This function translates a vector of 2-digit ISCO88COM codes to the ESEG class schema.

Usage

isco08_to_eseg(
  x,
  work_status,
  main_activity,
  age,
  type,
  label = FALSE,
  to_factor = FALSE
)

Arguments

x

A character vector of 2-digit ISCO codes. This should be the 4-digit equivalent so instead of 13, the code should be 1300, which is the 4-digit version of of the 2-digit ISCO.

work_status

A numeric vector of values from 0 to 3 where 1 = self_employed, 0 = employee and 2 = non employed.

main_activity

A numeric vector of values from 1 to 5 where 1 = respondent is working, 2 = respondent is in education, 3 = respondent is disabled , 4 = respondent has no paid work (household work, taking care of children, etc..) and 5 = respondent is retired. For an example, see the variable mainact from the European Social Survey.

age

A numeric vector of ages of the respondent.

type

The type of translation to make. Possible values are "one-digit" and "two-digit". The "one-digit" translation returns a broad summary based translation of only 9 categories, whereas the "two-digit" translation returns a much bigger ESEG translation of more than 25 categories.

label

A logical value indicating whether to return the labels of the translated ESEG codes (default is FALSE).

to_factor

A logical value indicating whether to return a factor instead of a character. The order of the labels is taken from the sorted codes of ESEG which can be found in the source code of each function.

Details

The translation was implemented following the work of Kea Tijdens in the document "ESEG-2014 coding scheme + explanatory note". For more details, see the last table in the document here.

For more details on this class schema, please check the references below:

Börlin, S., & Zettl, L. (2020). Mikrozensus Tools: Die Klassifikation European Socio-economic Groups (ESeG) im Mikrozensus Scientific Use File. (GESIS Papers, 2020/08). Köln: GESIS - Leibniz-Institut für Sozialwissenschaften. https://doi.org/10.21241/ssoar.68449

Bohr, J. (2018). EU-AES Tools: Implementation of the European Socioeconomic Groups Classification (ESeG) using Adult Education Survey Microdata. (GESIS Papers, 2018/14). Köln: GESIS - Leibniz-Institut für Sozialwissenschaften.https://doi.org/10.21241/ssoar.57622

Meron M, and all ESSnet members (2014) ESSnet ESeG Final Report. Paris, INSEE,Direction des Statistiques Démographiques et Sociales ESSnet project

Meron, M. et al. (2014): Final Report of the ESSnet on the harmonisation and implementation of a European socio-economic classification: European Socio-economic Groups (ESeG)

Tijdens, K.G. (2016) ESEG-2014 coding scheme + explanatory note. Deliverable 8.13 of the SERISS project funded under the European Union’s Horizon 2020 research and innovation programme GA No: 654221. Available at: https://seriss.eu/resources/deliverables

Resource websites of the European Socio-economic Groups (ESeG):

https://cros-legacy.ec.europa.eu/content/eseg_en

https://ec.europa.eu/eurostat/cros/content/eseg-report-technicaldocuments_en

https://ec.europa.eu/eurostat/cros/system/files/ESEG-Report-TechnicalAnnexes_0.zip_en

Examples

library(dplyr)

# Convert isco08 to two digits
ess$isco08_two <- isco08_swap(ess$isco08, from = 4, to = 2)

# Using the two-digit translation
ess %>%
  transmute(
    isco08_two,
    eseg = isco08_to_eseg(
      isco08_two,
      work_status,
      main_activity,
      agea,
      type = "two-digit"
    ),
    eseg_label = isco08_to_eseg(
      isco08_two,
      work_status,
      main_activity,
      agea,
      type = "two-digit",
      label = TRUE
    )
  )
#> # A tibble: 48,285 × 3
#>    isco08_two eseg  eseg_label                                                  
#>    <chr>      <chr> <chr>                                                       
#>  1 5400       4.2   42 Technicians, clerical support, services and sales self-e…
#>  2 1300       1.3   13 Higher managerial employees                              
#>  3 3100       3.1   31 Science, engineering and ICT technicians and associated …
#>  4 7100       6.1   61 Building and related trade employees                     
#>  5 6100       7.4   74 Agricultural employees                                   
#>  6 6100       7.4   74 Agricultural employees                                   
#>  7 9300       NA    NA                                                          
#>  8 1300       1.3   13 Higher managerial employees                              
#>  9 1300       1.1   11 Higher managerial self-employed                          
#> 10 6100       7.4   74 Agricultural employees                                   
#> # ℹ 48,275 more rows

# Using the one-digit translation
ess %>%
  transmute(
    isco08_two,
    eseg = isco08_to_eseg(
      isco08_two,
      work_status,
      main_activity,
      agea,
      type = "one-digit"
    ),
    eseg_label = isco08_to_eseg(
      isco08_two,
      work_status,
      main_activity,
      agea,
      type = "one-digit",
      label = TRUE
    )
  )
#> # A tibble: 48,285 × 3
#>    isco08_two eseg  eseg_label                                          
#>    <chr>      <chr> <chr>                                               
#>  1 5400       4     4 Small entrepreneur                                
#>  2 1300       1     1 Manager                                           
#>  3 3100       3     3 Technicians and associated professeional employees
#>  4 7100       6     6 Industrial skilled employees                      
#>  5 6100       7     7 Less skilled employees                            
#>  6 6100       7     7 Less skilled employees                            
#>  7 9300       NA    NA                                                  
#>  8 1300       1     1 Manager                                           
#>  9 1300       1     1 Manager                                           
#> 10 6100       7     7 Less skilled employees                            
#> # ℹ 48,275 more rows