Translate 3-digit ISCO08/ISCO88COM to MSEC

This function translates a vector of 3-digit ISCO08/ISCO88COM codes to MSEC codes using the translation tables stored in all_schema$isco08_to_msec / all_schema$isco88com_to_msec.

Usage

isco08_to_msec(x, is_supervisor, self_employed, n_employees, label = FALSE)

isco88com_to_msec(x, is_supervisor, self_employed, n_employees, label = FALSE)

Arguments

x

A character vector of 3-digit ISCO codes. This should be the 4-digit equivalent so instead of 131, the code should be 1310, which is the 4-digit version of of the 3-digit ISCO.

is_supervisor

A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).

self_employed

A numeric vector indicating whether each individual is self-employed (1) or an employee (0).

n_employees

A numeric vector indicating the number of employees under each respondent.

label

A logical value indicating whether to return the labels of the translated MSEC codes (default is FALSE).

Value

A character vector of MSEC codes.

Details

These translations were created from the document "Allocation rules of ISCO-08 and ISCO-88 (COM) 3-digit codes to ESEG-Revised" from Oscar Smallenbroek, Florian Hertel and Carlo Barone. For more info, please contact the authors. Although originally called ESEG-Revised, the class schema has been formally called MSEC.

This function will accept 3 digit codes as 4 digits. This means that if the 3-digit code is 131 then it should be 1310. All codes should be 4 digits, even though the code is represented as 3-digits (1310, 1320, etc..)

Examples

library(dplyr)

# convert to three digits
ess$isco08_three <- isco08_swap(ess$isco08, from = 4, to = 3)
ess$isco88com_three <- isco88_swap(ess$isco88com, from = 4, to = 3)

# isco08
ess %>%
  transmute(
    isco08_three,
    msec_label = isco08_to_msec(
      isco08_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    ),
    msec = isco08_to_msec(
      isco08_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    )
  )
#> # A tibble: 56,752 × 3
#>    isco08_three msec_label                                      msec 
#>    <chr>        <chr>                                           <chr>
#>  1 3110         Lower Management Employees                      14   
#>  2 4120         General Clerks                                  51   
#>  3 1120         Higher Management Employees                     13   
#>  4 2330         Primary and Secondary Education Professionals   25   
#>  5 7410         Skilled Labor Supervisors                       60   
#>  6 5310         Personal Care Employees                         53   
#>  7 NA           NA                                              NA   
#>  8 2430         Higher Management Employees                     13   
#>  9 2630         Professionals N.E.C. (Not Elsewhere Classified) 20   
#> 10 2640         Social Professionals                            24   
#> # ℹ 56,742 more rows

# isco88com
ess %>%
  transmute(
    isco88com_three,
    msec_label = isco88com_to_msec(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    ),
    msec = isco88com_to_msec(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    )
  )
#> # A tibble: 56,752 × 3
#>    isco88com_three msec_label                                    msec 
#>    <chr>           <chr>                                         <chr>
#>  1 3110            Lower Management Employees                    14   
#>  2 3430            General Clerks                                51   
#>  3 1210            Higher Management Employees                   13   
#>  4 2320            Primary and Secondary Education Professionals 25   
#>  5 7130            Skilled Labor Supervisors                     60   
#>  6 5130            Personal Care Employees                       53   
#>  7 NA              NA                                            NA   
#>  8 2410            Higher Management Employees                   13   
#>  9 3460            Legal and Social Associate Professionals      34   
#> 10 2440            Social Professionals                          24   
#> # ℹ 56,742 more rows