Translate ISCO68 to EGP

This function translates a vector of ISCO68 codes to EGP codes using the translation table stored in the all_schemas$isco68_to_egp data frame.

Usage

isco68_to_egp(x, self_employed, n_employees, label = FALSE)

Arguments

x

A character vector of ISCO68 codes.

self_employed

A numeric vector indicating whether each individual is self-employed (1) or not (0).

n_employees

A numeric vector indicating the number of employees for each individual.

label

A logical value indicating whether to return the labels of the translated EGP codes (default is FALSE).

Value

A character vector of EGP codes.

Examples

library(dplyr)

ess %>% mutate(EGP = isco68_to_egp(isco68, self_employed, emplno, label = TRUE))
#> # A tibble: 56,752 × 7
#>    isco68 isco88 isco08 emplno self_employed is_supervisor EGP                  
#>    <chr>  <chr>  <chr>   <dbl>         <dbl>         <dbl> <chr>                
#>  1 0140   3111   3111        0             0             1 'lo controllers'     
#>  2 3210   3431   4120        0             0             0 'routine nonmanual'  
#>  3 2119   1210   1120        0             0             1 'higher controllers' 
#>  4 1320   2320   2330        0             0             0 'lo controllers'     
#>  5 8550   7137   7411        0             0             1 'skilled manual'     
#>  6 5401   5131   5311        0             0             0 'semi-unskilld manua…
#>  7 NA     NA     NA          0             0             0 NA                   
#>  8 1593   2419   2432        0             0             1 'lo controllers'     
#>  9 1930   3460   2635        0             0             0 'lo controllers'     
#> 10 1950   2444   2643        0             0             0 'lo controllers'     
#> # ℹ 56,742 more rows
ess %>% mutate(EGP = isco68_to_egp(isco68, self_employed, emplno, label = FALSE))
#> # A tibble: 56,752 × 7
#>    isco68 isco88 isco08 emplno self_employed is_supervisor EGP  
#>    <chr>  <chr>  <chr>   <dbl>         <dbl>         <dbl> <chr>
#>  1 0140   3111   3111        0             0             1 2    
#>  2 3210   3431   4120        0             0             0 3    
#>  3 2119   1210   1120        0             0             1 1    
#>  4 1320   2320   2330        0             0             0 2    
#>  5 8550   7137   7411        0             0             1 8    
#>  6 5401   5131   5311        0             0             0 9    
#>  7 NA     NA     NA          0             0             0 NA   
#>  8 1593   2419   2432        0             0             1 2    
#>  9 1930   3460   2635        0             0             0 2    
#> 10 1950   2444   2643        0             0             0 2    
#> # ℹ 56,742 more rows