Translate 3-digit ISCO88COM to ESEC

This function translates a vector of 3-digit ISCO88COM codes to ESEC codes using the translation table stored in the all_schemas$isco88com_to_esec_three data frame. Note that in all_schemas$isco88com_to_esec_three the column ESEC refers to the simplified ESEC, which matches ISCO codes to ESEC codes directly, instead of using information on number of employees, self-employed, etc..

Usage

isco88com_to_esec(
  x,
  is_supervisor,
  self_employed,
  n_employees,
  full_method = TRUE,
  label = FALSE
)

Arguments

x

A character vector of 3-digit ISCO88COM codes. Even though these should be 3-digit, instead of 130, the code should be 1300, which is the 3-digit version of ISCO.

is_supervisor

A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).

self_employed

A numeric vector indicating whether each individual is self-employed (1) or not (0).

n_employees

A numeric vector indicating the number of employees for each individual.

full_method

logical value indicating whether to do the translation using the full method (uses ISCO codes as well as information on number of employees, self-employed, etc..) or the simplified method (matches ISCO codes directly with ESEC codes).

label

A logical value indicating whether to return the labels of the translated ESEC codes (default is FALSE).

Value

A character vector of ESEC codes.

Details

The codification used in the User's guide suggests that contrary to the full method, which uses whether the respondent is a supervisor, self-employed and whether the person has subordinates, the simple method matches directly the ISCO code to an ESEC code. For more info, please see page 17 of the European Socio-economic Classification (ESeC) User Guide (2006) by Rode, D. and Harrison, E.

This translation is borrowed from the iscogen Stata package. For more info, search for 'ISCO-88 -> ESEC' in the documentation of the iscogen package. If you have ISCO88, you can translate it to ISCO88COM using the function isco88_to_isco88com before translating to ESEC.

This function will accept 3-digit codes as 4 digits. This means that if the 3-digit code is 131 then it should be 1310. All codes should be 4 digits, even though the code is represented as 3 digits (1310, 1230, etc..)

Examples

library(dplyr)

# convert to three digits
ess$isco88com <- isco88_to_isco88com(ess$isco88)
ess$isco88com_three <- isco88_swap(ess$isco88com, from = 4, to = 3)

# Using the full method
ess %>%
  transmute(
    esec_label = isco88com_to_esec(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE
    ),
    esec = isco88com_to_esec(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE
    )
  )
#> # A tibble: 56,752 × 2
#>    esec_label                                                 esec 
#>    <chr>                                                      <chr>
#>  1 'Lower mgrs/professionals, higher supervisory/technicians' 2    
#>  2 'Intermediate occupations'                                 3    
#>  3 'Large employers, higher mgrs/professionals'               1    
#>  4 'Lower mgrs/professionals, higher supervisory/technicians' 2    
#>  5 'Lower supervisors and technicians'                        6    
#>  6 'Lower sales and service'                                  7    
#>  7 NA                                                         NA   
#>  8 'Large employers, higher mgrs/professionals'               1    
#>  9 'Intermediate occupations'                                 3    
#> 10 'Lower mgrs/professionals, higher supervisory/technicians' 2    
#> # ℹ 56,742 more rows

# Using the simple method
ess %>%
  transmute(
    esec_simple = isco88com_to_esec(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      label = FALSE,
      full_method = FALSE
    ),
    esec_simple_label = isco88com_to_esec(
      isco88com_three,
      is_supervisor,
      self_employed,
      emplno,
      label = TRUE,
      full_method = FALSE
    )
  )
#> # A tibble: 56,752 × 2
#>    esec_simple esec_simple_label                                         
#>    <chr>       <chr>                                                     
#>  1 2           'Lower mgrs/professionals, higher supervisory/technicians'
#>  2 3           'Intermediate occupations'                                
#>  3 1           'Large employers, higher mgrs/professionals'              
#>  4 2           'Lower mgrs/professionals, higher supervisory/technicians'
#>  5 8           'Lower technical'                                         
#>  6 7           'Lower sales and service'                                 
#>  7 NA          NA                                                        
#>  8 1           'Large employers, higher mgrs/professionals'              
#>  9 3           'Intermediate occupations'                                
#> 10 2           'Lower mgrs/professionals, higher supervisory/technicians'
#> # ℹ 56,742 more rows