Translate 4-digit ISCO08/ISCO88/ISCO68 to SIOPS

This function translates a vector of 4-digit ISCO08/ISCO88/ISCO68 codes to SIOPS codes using the translation tables stored in all_schema$isco08_to_siops / all_schema$isco88_to_siops / all_schema$isco68_to_siops.

Usage

isco08_to_siops(x, to_factor = FALSE)

isco88_to_siops(x, to_factor = FALSE)

isco68_to_siops(x, to_factor = FALSE)

Arguments

x

A character vector of 4-digit ISCO codes.

to_factor

A logical value indicating whether to return a factor instead of a character. The order of the labels is taken from the sorted codes of SIOPS in all_schemas$isco08_to_siops, all_schemas$isco88_to_siops and all_schemas$isco68_to_siops.

Value

A character vector of SIOPS codes.

Details

This translation was taken from the iscogen Stata package. For more details, check out the package documentation and search for ISCO08/ISCO88/ISCO88 -> SIOPS.

Since SIOPS doesn't have any labels, the labels is not availabe in this function.

For more information on this class schema, please check the references below:

Ganzeboom, H.B.G., P.M. De Graaf, D.J. Treiman (1992) A Standard International Socio-Economic Index of Occupational Status. Social Science Research 21: 1-56.

Ganzeboom, H.B.G., D.J. Treiman (1996) Internationally Comparable Measures of Occupational Status for the 1988 International Standard Classification of Occupations. Social Science Research 25: 201-239.

Examples

library(dplyr)

ess %>%
  transmute(
    isco08,
    isco88,
    isco68,
    siops_08 = isco08_to_siops(isco08),
    siops_88 = isco88_to_siops(isco88),
    siops_68 = isco68_to_siops(isco68)
  )
#> # A tibble: 48,285 × 6
#>    isco08 isco88 isco68 siops_08 siops_88 siops_68
#>    <chr>  <chr>  <chr>  <chr>    <chr>    <chr>   
#>  1 5414   5169   5890   27.89    30       30      
#>  2 1321   1222   2120   64.00    60       64      
#>  3 3135   8120   7200   34.91    40       38      
#>  4 7131   7141   9310   30.92    31       31      
#>  5 6111   6111   6220   43.53    40       21      
#>  6 6111   6111   6220   43.53    40       21      
#>  7 9313   9313   9595   15.00    15       15      
#>  8 1311   1221   6000   61.15    60       48      
#>  9 1311   1221   6000   61.15    60       48      
#> 10 6111   6111   6220   43.53    40       21      
#> # ℹ 48,275 more rows