Translates 4-digit ISCO88COM to WRIGHT.

This function translates a vector of 4-digit ISCO88COM codes to the E.O Wright class schema.

Usage

isco88com_to_wright(
  x,
  is_supervisor,
  self_employed,
  n_employees,
  control_work,
  control_daily,
  type,
  label = FALSE
)

Arguments

x

A character vector of 4-digit ISCO codes. This should be the 4-digit equivalent so instead of 131, the code should be 1310, which is the 4-digit version of of the 4-digit ISCO.

is_supervisor

A numeric vector indicating whether each individual is a supervisor (1, e.g. responsible for other employees) or not (0).

self_employed

A numeric vector indicating whether each individual is self-employed (1) or an employee (0).

n_employees

A numeric vector indicating the number of employees under each respondent.

control_work

A likert-scale type question from 0 to 10 where 0 is whether an individual has no control over their work/organisation decisions and 10 is complete control over work/organization decisions. For an example, see the variable iorgact in the European Social Survey.

control_daily

A likert-scale type question from 1 to 4 where 1 means complete control to decide how their own daily work is/was organised and 4 means no control to decide how their own daily work is/was organised. For an example, see the variable orgwrk in the European Social Survey. Another example is recoding the variable wkdcorga from the European Social Survey such that 8-10 is 1, 5-7 is 2, 2-4 is 3 and 0-1 is 4.

type

The type of translation to make. Possible values are "simple", "decision-making" and "power-class".

label

A logical value indicating whether to return the labels of the translated WRIGHT codes (default is FALSE).

Details

The translation implemented in this function was originally developed by Erik Olin Wright. There are three possible types of translations: the "simple" version, the "decision-making" version and the "power-class" version. This translation was implemented following the SPSS script from Håkon Leiulfsrud and Heidi Jensberg. For more information, please contact the authors.

Examples

library(dplyr)

# E.O Wright - Simple translation
ess %>%
  transmute(
    isco88com,
    wr_simple = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "simple"
    ),
    wr_simple_label = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "simple",
      label = TRUE
    )
  )
#> # A tibble: 56,752 × 3
#>    isco88com wr_simple wr_simple_label        
#>    <chr>     <chr>     <chr>                  
#>  1 3111      7         Skilled manager/superv 
#>  2 3431      12        Low skilled workers    
#>  3 1210      4         Expert managers        
#>  4 2320      9         Skilled workers        
#>  5 7137      7         Skilled manager/superv 
#>  6 5131      12        Low skilled workers    
#>  7 NA        3         Self empl w/no empoyees
#>  8 2419      4         Expert managers        
#>  9 3460      9         Skilled workers        
#> 10 2444      6         Expert workers         
#> # ℹ 56,742 more rows

# E.O Wright - Decision-making translation
ess %>%
  transmute(
    isco88com,
    wr_decision = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "decision-making"
    ),
    wr_decision_label = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "decision-making",
      label = TRUE
    )
  )
#> # A tibble: 56,752 × 3
#>    isco88com wr_decision wr_decision_label      
#>    <chr>     <chr>       <chr>                  
#>  1 3111      9           Skilled workers        
#>  2 3431      12          Low skilled workers    
#>  3 1210      NA          NA                     
#>  4 2320      NA          NA                     
#>  5 7137      8           Skilled supervisors    
#>  6 5131      12          Low skilled workers    
#>  7 NA        3           Self empl w/no empoyees
#>  8 2419      5           Expert supervisors     
#>  9 3460      NA          NA                     
#> 10 2444      6           Experts                
#> # ℹ 56,742 more rows

# E.O Wright - Power-class translation
ess %>%
  transmute(
    isco88com,
    wr_power = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "power-class"
    ),
    wr_power_label = isco88com_to_wright(
      isco88com,
      is_supervisor,
      self_employed,
      emplno,
      control_work,
      control_daily,
      type = "power-class",
      label = TRUE
    )
  )
#> # A tibble: 56,752 × 3
#>    isco88com wr_power wr_power_label             
#>    <chr>     <chr>    <chr>                      
#>  1 3111      7        Skilled workers            
#>  2 3431      6.5      Low skilled Semi autonomous
#>  3 1210      NA       NA                         
#>  4 2320      NA       NA                         
#>  5 7137      5        Skilled supervisors        
#>  6 5131      7.5      Low skilled workers        
#>  7 NA        3        Self empl                  
#>  8 2419      5        Skilled supervisors        
#>  9 3460      NA       NA                         
#> 10 2444      7        Skilled workers            
#> # ℹ 56,742 more rows